# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Guide Soft IT Solutions API
  description: API for Guide Soft IT Solutions platform including courses, enrollments, services, and payments
  version: 1.0.0
  contact:
    name: Guide Soft Support
    email: support@guidesoft.com

servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.guidesoft.com/api
    description: Production server

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                  minLength: 2
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
        '500':
          description: Internal server error

  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error

  /auth/logout:
    post:
      summary: Logout user
      tags:
        - Authentication
      responses:
        '200':
          description: Logout successful
        '500':
          description: Internal server error

  /auth/me:
    get:
      summary: Get current user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
        '404':
          description: User not found
        '500':
          description: Internal server error

  /courses:
    get:
      summary: List all courses
      tags:
        - Courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Courses retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Internal server error

    post:
      summary: Create a new course
      tags:
        - Courses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          description: Not authenticated
        '500':
          description: Internal server error

  /courses/{id}:
    get:
      summary: Get course by ID
      tags:
        - Courses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '404':
          description: Course not found
        '500':
          description: Internal server error

    put:
      summary: Update course
      tags:
        - Courses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          description: Not authenticated
        '404':
          description: Course not found
        '500':
          description: Internal server error

    delete:
      summary: Delete course
      tags:
        - Courses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course deleted
        '401':
          description: Not authenticated
        '404':
          description: Course not found
        '500':
          description: Internal server error

  /enrollments:
    get:
      summary: List user enrollments
      tags:
        - Enrollments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Enrollments retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enrollment'
        '401':
          description: Not authenticated
        '500':
          description: Internal server error

    post:
      summary: Create enrollment
      tags:
        - Enrollments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseId
              properties:
                courseId:
                  type: string
      responses:
        '201':
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          description: Already enrolled
        '401':
          description: Not authenticated
        '500':
          description: Internal server error

  /services:
    get:
      summary: List all services
      tags:
        - Services
      responses:
        '200':
          description: Services retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          description: Internal server error

  /services/booking:
    post:
      summary: Book a service
      tags:
        - Services
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - serviceId
                - date
                - time
              properties:
                serviceId:
                  type: integer
                date:
                  type: string
                  format: date
                time:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid input
        '401':
          description: Not authenticated
        '500':
          description: Internal server error

  /payments/create-checkout:
    post:
      summary: Create Stripe checkout session
      tags:
        - Payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseId
                - amount
              properties:
                courseId:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
        '401':
          description: Not authenticated
        '500':
          description: Internal server error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        avatar:
          type: string
        bio:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time

    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        level:
          type: string
        duration:
          type: string
        lessons:
          type: integer
        rating:
          type: number
        students:
          type: integer
        image:
          type: string
        featured:
          type: boolean
        instructorId:
          type: string
        instructor:
          $ref: '#/components/schemas/Instructor'

    CourseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        level:
          type: string
        duration:
          type: string
        lessons:
          type: integer
        image:
          type: string
        featured:
          type: boolean
        instructorId:
          type: string

    CourseDetail:
      allOf:
        - $ref: '#/components/schemas/Course'
        - type: object
          properties:
            lessonsList:
              type: array
              items:
                $ref: '#/components/schemas/Lesson'
            enrollments:
              type: array
              items:
                $ref: '#/components/schemas/Enrollment'

    Instructor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        bio:
          type: string

    Lesson:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: string
        duration:
          type: integer
        order:
          type: integer

    Enrollment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        courseId:
          type: string
        progress:
          type: integer
        completed:
          type: boolean
        enrolledAt:
          type: string
          format: date-time
        course:
          $ref: '#/components/schemas/Course'

    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        price:
          type: string
        features:
          type: array
          items:
            type: string

    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceId:
          type: integer
        date:
          type: string
          format: date-time
        time:
          type: string
        notes:
          type: string
        status:
          type: string

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer